Big refactoring of gigantic matchMetaDataFromFiles controller method
